requirements.txt
streamlit==1.24.0
opencv-python==4.7.0.72
numpy==1.24.3
matplotlib==3.7.1
moviepy==1.0.3
Pillow==9.5.0